+++ src/decorators/audit-log.decorator.ts
@@ -1,0 +1,22 @@
+import { applyDecorators, SetMetadata, UseInterceptors } from "@nestjs/common";
+
+import { LoggingInterceptor } from "../interceptors";
+import { AuditLogMetadata } from "../interfaces";
+
+const METADATA_KEY = 'auditLog';
+
+export function AuditLog({
+  label = '',
+  omitFields = [],
+  requestBodyOutFields = [],
+  requestParamsOutFields = [],
+  requestQueryOutFields = [],
+  responseBodyOutFields = [],
+}: AuditLogMetadata = {}) {
+  return applyDecorators(
+    SetMetadata(METADATA_KEY, { label, omitFields, requestBodyOutFields, requestParamsOutFields, requestQueryOutFields, responseBodyOutFields }),
+    UseInterceptors(LoggingInterceptor),
+  );
+}
+
+export default AuditLog;