+++ src/interfaces/index.ts
@@ -1,3 +1,4 @@
+export { AuditLogMetadata } from "./reflector-metadata.interface";
 export { UserPermissionSet, CurrentUser } from "./current-user.interface";
 export { ValidatorParams } from "./lookup-type.interface";
 export {