+++ src/interfaces/reflector-metadata.interface.ts
@@ -1,0 +1,8 @@
+export interface AuditLogMetadata {
+  label?: string;
+  omitFields?: string[];
+  requestBodyOutFields?: string[] | '*' | 'all';
+  requestParamsOutFields?: string[] | '*' | 'all';
+  requestQueryOutFields?: string[] | '*' | 'all';
+  responseBodyOutFields?: string[] | '*' | 'all';
+}