+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+## [17.8.2](https://github.com/US-EPA-CAMD/easey-common/compare/v17.8.1...v17.8.2) (2024-11-15)
+
+
+### Bug Fixes
+
+* moreInfo updated ([422febd](https://github.com/US-EPA-CAMD/easey-common/commit/422febdcee1f75d23effbb4df44d8b97451a72c3))
+* update the filterRequestparam condition ([93fa6e8](https://github.com/US-EPA-CAMD/easey-common/commit/93fa6e86ecaaf5080df5e8abb8c6136dbdd19c3a))
+* updated with inFields and query,body,params ([efb8fe3](https://github.com/US-EPA-CAMD/easey-common/commit/efb8fe37053f23b05c55343e2ec3b47726cbd7b0))
+
+## [17.8.1](https://github.com/US-EPA-CAMD/easey-common/compare/v17.8.0...v17.8.1) (2024-11-13)
+
+
+### Bug Fixes
+
+* fix ip address and json stringify message ([b07d866](https://github.com/US-EPA-CAMD/easey-common/commit/b07d866f68aa0e1350a3254c824cd723a2c902ef))
+
+# [17.8.0](https://github.com/US-EPA-CAMD/easey-common/compare/v17.7.4...v17.8.0) (2024-11-06)
+
+
+### Features
+
+* **6246:** added controller classname to logging interceptor ([e9a8df9](https://github.com/US-EPA-CAMD/easey-common/commit/e9a8df95de67574f44b6f57f671f9bf216bb7f02))
+* **6426:** added AuditLog wrapper decorator ([c03bf3a](https://github.com/US-EPA-CAMD/easey-common/commit/c03bf3a34c9937d75ddc6456ffa6a4aee596415e))
+
 ## [17.7.4](https://github.com/US-EPA-CAMD/easey-common/compare/v17.7.3...v17.7.4) (2024-10-23)